miatoll: Use Headers from our kernel build tree
This check is to pick the kernel headers from the right location.
If the macro above is defined, we make the assumption that we have the kernel
available in the build tree.
If the macro is not present, the headers are picked from hardware/qcom/msmXXXX
failing which, they are picked from bionic.

REQUIRED to copy headers require for Audio and Display hal

Signed-off-by: Arjun-Ingole <arjuningole38@gmail.com>